re-organized code structure, added products categories and reviews data into the context
