+++ resolvers/Category.js
@@ -1,0 +1,7 @@
+exports.Category = {
+    products: (parent, args, { products }) => {
+        const { id } = parent;
+
+        return products.filter((product) => product.categoryId === id);
+    },
+};
