+++ resolvers/Product.js
@@ -1,0 +1,7 @@
+exports.Product = {
+    category: (parent, args, { categories }) => {
+        const { categoryId } = parent;
+
+        return categories.find((category) => category.id === categoryId);
+    },
+};
